@@ -28,22 +28,23 @@
     "url": "https://github.com/BaekjoonHub/BaekjoonHub.git"
   },
   "devDependencies": {
+    "@eslint/js": "^9.33.0",
+    "@types/chrome": "^0.1.1",
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^13.0.0",
-    "eslint": "^8.57.1",
-    "eslint-config-prettier": "^10.1.5",
+    "eslint": "^9.33.0",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-prettier": "^5.5.1",
-    "globals": "^16.0.0",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-prettier": "^5.5.4",
+    "globals": "^16.3.0",
     "prettier": "^3.6.2",
     "rimraf": "^6.0.1",
     "selenium-webdriver": "^4.34.0",
-    "webpack": "^5.98.0",
+    "webpack": "^5.101.0",
     "webpack-cli": "^6.0.1"
   },
   "dependencies": {
-    "chrome": "^0.1.0",
     "fflate": "^0.8.2",
     "file-saver": "^2.0.5",
     "js-sha1": "^0.6.0",
