@@ -1,8 +1,8 @@
-module.exports = {
+export default {
   singleQuote: false,
   semi: true,
   useTabs: false,
   tabWidth: 2,
-  trailingComma: 'es5',
+  trailingComma: "es5",
   printWidth: 200,
 };
